Docker/yahoo_llm_project/scripts/requirements.txt
psycopg2-binary==2.9.9
bert-score==0.3.13
torch>=2.0.0
transformers>=4.30.0
